Remove macros module
