+++ serde_arrow/src/test_impls/mod.rs
@@ -5,7 +5,6 @@ mod dictionary;
 mod examples;
 mod json_values;
 mod list;
-mod macros;
 mod map;
 mod primitives;
 mod r#struct;
